+++ CHANGELOG.md
@@ -427,3 +427,5 @@
 * Removed duplicate ChangeSceneSpecial
 * DoorInteractableComponent and ChangeSceneSpecial can now spawn a transfer effect when used
 * Volume control is now pseudo-logarithmic (exponential)
+* Weapon viewmodels are now explicitly initialized (WeaponViewModelScript.Init)
+* Sprite weapon viewmodels can be (pseudo) attached to ViewModel camera